Don’t ever initialize a primary CA within a secondary DC
