+++ agent/consul/server.go
@@ -142,12 +142,6 @@ type Server struct {
 	caProviderRoot *structs.CARoot
 	caProviderLock sync.RWMutex
 
-	// caPruningCh is used to shut down the CA root pruning goroutine when we
-	// lose leadership.
-	caPruningCh      chan struct{}
-	caPruningLock    sync.RWMutex
-	caPruningEnabled bool
-
 	// Consul configuration
 	config *Config
